% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce_markov.R
\name{ce_markov}
\alias{ce_markov}
\title{This is an example function}
\usage{
ce_markov(
  start_pop,
  p_matrix,
  state_c_matrix,
  trans_c_matrix,
  state_q_matrix,
  n_cycles = 46,
  init_age = 55,
  s_names = NULL,
  t_names = NULL
)
}
\arguments{
\item{start_pop}{This is the starting population for the model}

\item{t_names}{}
}
\value{
this is what the function gives you back. e.g. total costs and total QALYs
}
\description{
This is an example function
}
